Refactor zim_manager
Add "contract" to zim_manager, and the contained packages: fileselect_view and library_view
