+++ app/src/main/java/org/kiwix/kiwixmobile/zim_manager/ZimManagePresenter.java
@@ -26,6 +26,7 @@ import org.kiwix.kiwixmobile.R;
 import org.kiwix.kiwixmobile.base.BasePresenter;
 import org.kiwix.kiwixmobile.downloader.DownloadService;
 import org.kiwix.kiwixmobile.utils.SharedPreferenceUtil;
+import org.kiwix.kiwixmobile.zim_manager.contract.ZimManageViewCallback;
 
 import javax.inject.Inject;
