+++ app/src/main/java/org/kiwix/kiwixmobile/zim_manager/fileselect_view/contract/ZimFileSelectViewCallback.java
@@ -15,31 +15,16 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.zim_manager.library_view;
+package org.kiwix.kiwixmobile.zim_manager.fileselect_view.contract;
 
 import org.kiwix.kiwixmobile.base.ViewCallback;
 import org.kiwix.kiwixmobile.library.entity.LibraryNetworkEntity;
-import org.kiwix.kiwixmobile.library.entity.LibraryNetworkEntity.Book;
 
-import java.util.LinkedList;
+import java.util.ArrayList;
 
 /**
  * Created by EladKeyshawn on 06/04/2017.
  */
-
-public interface LibraryViewCallback extends ViewCallback {
-
-  void showBooks(LinkedList<Book> books);
-
-  void displayNoNetworkConnection();
-
-  void displayNoItemsFound();
-
-  void displayNoItemsAvailable();
-
-  void displayScanningContent();
-
-  void stopScanningContent();
-
-  void downloadFile(LibraryNetworkEntity.Book book);
+public interface ZimFileSelectViewCallback extends ViewCallback {
+  void showFiles(ArrayList<LibraryNetworkEntity.Book> books);
 }
