+++ app/src/main/java/org/kiwix/kiwixmobile/zim_manager/library_view/contract/LibraryViewCallback.java
@@ -15,13 +15,31 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.zim_manager;
+package org.kiwix.kiwixmobile.zim_manager.library_view.contract;
 
 import org.kiwix.kiwixmobile.base.ViewCallback;
+import org.kiwix.kiwixmobile.library.entity.LibraryNetworkEntity;
+import org.kiwix.kiwixmobile.library.entity.LibraryNetworkEntity.Book;
+
+import java.util.LinkedList;
 
 /**
- * Created by srv_twry on 15/2/18.
+ * Created by EladKeyshawn on 06/04/2017.
  */
 
-public interface ZimManageViewCallback extends ViewCallback {
+public interface LibraryViewCallback extends ViewCallback {
+
+  void showBooks(LinkedList<Book> books);
+
+  void displayNoNetworkConnection();
+
+  void displayNoItemsFound();
+
+  void displayNoItemsAvailable();
+
+  void displayScanningContent();
+
+  void stopScanningContent();
+
+  void downloadFile(LibraryNetworkEntity.Book book);
 }
